Move AbstractTaskCheduler and TaskScheduler* to tajo.querymaster package.
